fix(telecom-dashboard): lint scripts entries to fix package main entry
Main entry (CJS) is currently not generated because we launch build task in development mode.

CJS & UMD are only generated in production mode. Currently, if you import the package using NPM, main and browser enries are not availables.

Script entries are now linted as others packages, to avoid this issue.