fix: respect order in js loaders (#2947)
Before:

If sources files need to be compiled with Babel plugins and don't have a native native compliance with JS, Webpack is not able to compile the source, even if the right plugins to compile are defined.

This behavior is due to multiple loader for JS files: Babel / translation-inject / translation-ui-router. In fact, translation-inject and translation-ui-router try to parse and compile the code, without any compilation by Babel. The source code is not valid with JS compiler, this 2 loaders are crashing.

The solution consist to use babel-loader before translation-inject and translation-ui-router. Babel does the job, and the 2 following loaders are able to do their own.

But, another issue then arises, babel remove comment like @ngTranslationsInject and prevent translations loader to do thir job. We should let this commit when Babel is used.

After:

Babel-loader is used before translation-inject and translation-ui-router and let the comment like @ngTranslationsInject.

Webpack is able to compile with custom plugins which replace custom syntax that JS engine doesn't understand ;)

Signed-off-by: Joffrey LEVEUGLE <joffrey.leveugle@corp.ovh.com>
Co-authored-by: Antoine Leblanc <antoine.leblanc@corp.ovh.com>