fix(telephony.contact): add wayName directory field
As belgian customers can't update the address field due to API restriction, the address field is removed, for all customers. This way all customers (French, Belgians) will be able to update the wayName of their address. French customers would still be able to update their complete address with wayName and existing others fields. This choice is made to avoid using feature flipping for just one field, and also ng-if !== BE for disabling the address field.

ref: DTRSD-30994
Signed-off-by: JeremyDec <jeremy.de-cesare@ovhcloud.com>